添加了菜品管理、套餐管理。

+++ sky-server/src/main/java/com/sky/service/impl/CategoryServiceImpl.java
@@ -3,11 +3,17 @@ package com.sky.service.impl;
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.baomidou.mybatisplus.core.metadata.IPage;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
+import com.sky.constant.MessageConstant;
 import com.sky.constant.StatusConstant;
 import com.sky.dto.CategoryDTO;
 import com.sky.dto.CategoryPageQueryDTO;
 import com.sky.entity.Category;
+import com.sky.entity.Dish;
+import com.sky.entity.Setmeal;
+import com.sky.exception.DeletionNotAllowedException;
 import com.sky.mapper.CategoryMapper;
+import com.sky.mapper.DishMapper;
+import com.sky.mapper.SetmealMapper;
 import com.sky.result.PageResult;
 import com.sky.service.CategoryService;
 import org.springframework.beans.BeanUtils;
@@ -20,6 +26,13 @@ import java.util.List;
 public class CategoryServiceImpl implements CategoryService {
     @Autowired
     private CategoryMapper categoryMapper;
+
+    @Autowired
+    private SetmealMapper setmealMapper;
+
+    // TODO 我记得Spring中讲过创建Bean，分为创建一个Bean，和创建不同的Bean。复习一下！！！
+    @Autowired
+    private DishMapper dishMapper;
 
     public void save(CategoryDTO categoryDTO) {
         // 不用检查菜品分类的名字是否已存在，因为GlobalExceptionHandler中已经有SQL异常捕获方法来完成这个功能了，因为数据库中name设置的是唯一约束
@@ -38,6 +51,21 @@ public class CategoryServiceImpl implements CategoryService {
 
 
     public void deleteById(Long id) {
+        // TODO 这个地方，因为分类表被套餐表和菜品表外键指向，所以不能直接删除的
+        LambdaQueryWrapper<Setmeal> lqw1 = new LambdaQueryWrapper<>();
+        lqw1.eq(Setmeal::getCategoryId, id);
+        List<Setmeal> setmeals = setmealMapper.selectList(lqw1);
+        if (setmeals != null && setmeals.size() > 0) {
+            throw new DeletionNotAllowedException(MessageConstant.CATEGORY_BE_RELATED_BY_SETMEAL);
+        }
+
+        LambdaQueryWrapper<Dish> lqw2 = new LambdaQueryWrapper<>();
+        lqw2.eq(Dish::getCategoryId, id);
+        List<Dish> dishes = dishMapper.selectList(lqw2);
+        if (dishes != null && dishes.size() > 0) {
+            throw new DeletionNotAllowedException(MessageConstant.CATEGORY_BE_RELATED_BY_DISH);
+        }
+
         categoryMapper.deleteById(id);
     }
